# Conflicts:
#	src/main/java/com/open3hr/adeies/app/Auth/AuthenticationService.java
#	src/main/java/com/open3hr/adeies/app/employee/controller/EmployeeController.java
#	src/main/java/com/open3hr/adeies/app/user/service/impl/UserServiceImpl.java
#	src/main/java/com/open3hr/adeies/configuration/JwtConfigurationfilter.java

+++ src/main/java/com/open3hr/adeies/app/leaveRequest/service/impl/LeaveRequestServiceImpl.java
@@ -54,7 +54,10 @@ public class LeaveRequestServiceImpl implements LeaveRequestService {
 
     @Override
     public void deleteById(Integer id) {
+        Optional<LeaveRequest> leaveRequest = leaveRequestRepository.findById(id);
+        if (leaveRequest.isPresent() && leaveRequest.get().getStatus()==Status.PENDING)
         this.leaveRequestRepository.deleteById(id);
+        else throw new RuntimeException("oxi denn eee ");
     }
 
     @Override
